Write a Ceramic Class that accesses the Conviction State doc
This basic Ceramic class is able to initialized, read, and write back
the convictions state document.

I've attempted to implement a generic interface that would eventually
allow some pluggability for the storage layer.

The next step is implementing some test tooling to create the proposals
and conviction documents for users, so I can fill out the rest of the
class. I've considered creating a basic webUI to mock this state but I
think doing this via headless tests will yield more maintainability
gains in the long run.

Test Plan: None Yet; See above. I've implemented a basic test function
that interacts with ceramic when the ceramic.ts file is invoked, but that's not
sufficent, obviously. That was just for sanity checking to ensure this code
does something correctly.