Add treasury and, CORS, and proposerAllowList configs
The treasury config uses a given address' holdings as the requestable
funds pool. If this is not configured, like if no fundable proposals
will be submitted, the total supply is used as the requestable funds
pool for calculation purposes. This simplifies the calculation somewhat
but is only recommended for proposals that have a zero-amount.

CORS configs allow users to configure allowable endpoints in the event
that the service exists on a different endpoint from the frontend and no
proxy server is used to unite them. This is mostly for convenience.

the proposerAllowlist is a workaround for state document control at the
ceramic layer. It allows only proposals from specified address'
conviction docs to tbe added to the state doc.